+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Use Rust nightly for formatting (more options)
 - Use buster-slim base Docker image
 - Renaming admin commands with `admin-` suffix (admin-help, admin-enable, admin-sync)
+- Merge command results in one comment (if comments are sent)
 
 ## [0.9.1] - 2021-03-01